% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_category.R
\name{add_category}
\alias{add_category}
\title{Append universe with existing category}
\usage{
add_category(universe, category)
}
\arguments{
\item{universe}{Output object from \code{universe()}.}

\item{category}{Output object from \code{category()}.}
}
\value{
A \code{list} with special class \code{universe}.
}
\description{
This appends an existing \code{universe()} object with an existing \code{category()} object.
This does the same checks done in \code{universe()} but just on what is passed to \code{category()}.
}
\examples{
uni <- universe(
    data = mtcars,

    category(
        name = "vs",
        groups = c(0, 1),
        targets = c(1/2, 1/2)
    )
)

cat <- category(
    name = "am",
    groups = c(0, 1),
    targets = c(1/3, 2/3)
)

add_category(uni, cat)

}
