% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_weights.R
\name{append_weights}
\alias{append_weights}
\title{Append weights generated by iterake to original data}
\usage{
append_weights(iterake, wgt_name = "weight")
}
\arguments{
\item{iterake}{Output object created with \code{iterake()} function.}

\item{wgt_name}{Name given to column of weights to be added to data, default is "weight".}
}
\value{
A \code{tibble} consisting of the original data plus the column of weights.
}
\description{
This function appends the created weights from \code{iterake()} to the original data file that
was used in \code{universe()}.
}
