% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-universe.R
\name{autoplot.universe}
\alias{autoplot.universe}
\title{Visualize \code{universe()}.}
\usage{
\method{autoplot}{universe}(object, ...)
}
\arguments{
\item{object}{Output of \code{universe()}.}

\item{...}{Not currently used.}
}
\description{
This returns a plot comparing the actual unweighted marginal proportions with
the targets of the categories used in constructing the universe.
}
