% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_margin.R
\name{build_margin}
\alias{build_margin}
\title{Create weighting category}
\usage{
build_margin(name, buckets, targets, sum.1 = FALSE)
}
\arguments{
\item{name}{Name given to weighting category, character. Must have exact match in the column names of data you intend to weight.}

\item{buckets}{Vector corresponding to the "buckets" in which your sample is divided.}

\item{targets}{Numeric vector of the proportions each element of \code{`buckets`} represents in the population.}

\item{sum.1}{Whether or not to force inputs to \code{`targets`} to sum to one.}
}
\value{
A nested \code{tibble} with special class \code{build_margin}.
}
\description{
To be fed into \code{universe()}, this function creates an individual weighting category 
with known marginal probabilities. (E.g., age group.)
}
\examples{
build_margin(
  name = "costume",
  buckets = c("Bat Man", "Cactus"),
  targets = c(0.645, 0.355)
)

build_margin(
  name = "seeds",
  buckets = c("Tornado", "Bird", "Earthquake"),
  targets = c(0.3333, 0.3333, 0.3333),
  sum.1 = TRUE
)

}
