% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{category}
\alias{category}
\title{Create weighting category}
\usage{
category(name, buckets, targets, sum.1 = FALSE)
}
\arguments{
\item{name}{Name given to weighting category, character. 
Must have exact match in the column names of data to be weighted.}

\item{buckets}{Vector corresponding to the "buckets" in which your category can be classified.}

\item{targets}{Numeric vector of the proportions each element of \code{`buckets`} represents in the population.}

\item{sum.1}{Whether or not to force inputs to \code{`targets`} to sum to one.}
}
\value{
A nested \code{tibble} with special class \code{category}.
}
\description{
This function creates an individual weighting category with known marginal 
probabilities. (E.g., age group, eye color.) One or more of these are built and
fed into \code{universe()}.
}
\examples{
category(
  name = "costume",
  buckets = c("Bat Man", "Cactus"),
  targets = c(0.645, 0.355)
)

category(
  name = "seeds",
  buckets = c("Tornado", "Bird", "Earthquake"),
  targets = c(0.3333, 0.3333, 0.3333),
  sum.1 = TRUE
)

}
