% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_inherit.R
\name{category_inherit}
\alias{category_inherit}
\title{Create weighting category from data}
\usage{
category_inherit(name, df, prev.wgt)
}
\arguments{
\item{name}{Name given to weighting category, character. 
Must have exact match in the column names of data you intend to weight.}

\item{df}{Data frame containing data you intend to weight.}

\item{prev.wgt}{Variable name of weight used to calculate bucket proportions in \code{`df`}. Optional.}
}
\value{
A nested \code{tibble} with special class \code{category}.
}
\description{
This function creates an individual weighting category with known marginal 
probabilities (E.g., age group, eye color.), but unlike \code{category()} it 
creates them from a dataframe instead of being directly assigned. One or more 
of these are built and fed into \code{universe()}.
}
\examples{
data(weight_me)

category_inherit(
  name = "costume",
  df = dplyr::filter(weight_me, satisfied == "No")
)

category_inherit(
  name = "transport",
  df = dplyr::filter(weight_me, satisfied == "No"),
  prev.wgt = prev_weight
)

}
