% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_margins.R
\name{compare_margins}
\alias{compare_margins}
\title{Compare unweighted or weighted data to targets}
\usage{
compare_margins(universe, data, weight, plot = FALSE)
}
\arguments{
\item{universe}{Full weighting design created by \code{universe()}.}

\item{data}{Data frame that includes weight from \code{iterake()}, optional.}

\item{weight}{Name of weight variable in optional \code{`data`} object, optional.}

\item{plot}{Display plot, default = FALSE.}
}
\value{
A tibble of unweighted counts and proportions, difference between 
unweighted and target proportions. If \code{`data`} and \code{`weight`} is given, it also includes
weighted counts and proportions, difference between weighted and target proportions.
Optionally, a plot of this information.
}
\description{
Inspect by what degree the unweighted or weighted sample proportions deviate from the
target proportions.
}
\examples{
data(demo_data)

mod <- universe(
        data = demo_data,
        
        category(
            name = "Sex",
            buckets = factor(
                x = levels(demo_data[["Sex"]]),
                levels = levels(demo_data[["Sex"]])
            ),
            targets = c(0.4, 0.5),
            sum.1 = TRUE
        ),

        category(
            name = "BirthYear",
            buckets = c(1986:1990),
            targets = rep(0.2, times = 5)
        ),
    
        category(
            name = "EyeColor",
            buckets = c("brown", "green", "blue"),
            targets = c(0.8, 0.1, 0.1)
        ),
    
        category(
            name = "HomeOwner",
            buckets = c(TRUE, FALSE),
            targets = c(3/4, 1/4)
        )
    )

compare_margins(
    universe = mod,
    plot = TRUE
)

wgts <- iterake(
    universe = mod
)

compare_margins(
    universe = mod,
    data = wgts,
    weight = weight,
    plot = FALSE
)

}
