% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealer_data.R
\docType{data}
\name{dealer_data}
\alias{dealer_data}
\title{Simulated car dealership data}
\format{A data frame with 200 rows and six columns:
\describe{
   \item{Sales_ID}{unique sales record}
   \item{Age}{age group of customer}
   \item{Year}{year in which the sale occurred}
   \item{Type}{body style of the vehicle purchased}
}}
\usage{
data(dealer_data)
}
\description{
A sample of 400 observations from a customer satisfaction survey.
}
\examples{
data(dealer_data)
}
\keyword{datasets}
