% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Simulated demographics data.}
\format{
A data frame with 200 rows and five columns:
\describe{
   \item{ID}{unique sample identifier}
   \item{Sex}{sex of respondent, factor}
   \item{BirthYear}{birth year of respondent (spanning from 1986-1990), integer}
   \item{EyeColor}{eye color of respondent, character}
   \item{HomeOwner}{whether or not respondent is a homeowner, logical}
}
}
\usage{
data(demo_data)
}
\description{
Dataset containing the respondent identifier, sex, birth year, eye color,
and whether or not the respondent is a homeowner.
}
\keyword{datasets}
