% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{effective_ss}
\alias{effective_ss}
\title{Effective sample size.}
\usage{
effective_ss(x, w)
}
\arguments{
\item{x}{A vector.}

\item{w}{A vector of weights.}
}
\description{
Effective sample size.
}
\examples{
effective_ss(
    x = runif(3),
    w = c(0.75, 1.00, 1.25)
)

}
