% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weights.R
\name{extract_weights}
\alias{extract_weights}
\title{Extract weights}
\usage{
extract_weights(x, name = ".weight", append = TRUE)
}
\arguments{
\item{x}{Output of \code{iterake()}.}

\item{name}{Desired column name of the weights, default is \code{".weight"}.}

\item{append}{Whether or not to either append the weights to the original
data (\code{TRUE}, default), or to only return a \code{tibble} of the weights.}
}
\value{
A \code{tibble} of the weights, optionally with the original data.
}
\description{
Extracts the weights from the output of \code{iterake()}. It either returns
data passed to \code{universe()} with the weights appended, or just a one-column
\code{tibble} of the weights.
}
