% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HairEyeColor2.R
\docType{data}
\name{HairEyeColor2}
\alias{HairEyeColor2}
\title{Dummy data}
\format{
A tibble with 4 columns.
\itemize{
\item \code{id} - Unique ID
\item \code{hair} - A factor variable with 4 levels representing hair color - Black, Brown, Red, and Blond
\item \code{eye} - A factor variable with 4 levels representing eye color - Brown, Blue, Hazel, and Green
\item \code{sex} - A factor variable with 2 levels representing gender - Male and Female
}
}
\usage{
HairEyeColor2
}
\description{
This is just a \code{tibble} version of the \code{datasets::HairEyeColor} data,
then randomly sampled without replacement and added IDs.
}
\keyword{datasets}
