% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_category.R
\name{inherit_category}
\alias{inherit_category}
\title{Create weighting category from data}
\usage{
inherit_category(name, from, weight)
}
\arguments{
\item{name}{Name given to weighting category, character. 
Must have exact match in the column names of data you intend to weight as well as the
existing data from where targets are being pulled.}

\item{from}{Existing data frame from where targets will be pulled.}

\item{weight}{Optionally provide an existing weight variable to calculate 
weighted target proportions. Omit of unweighted target proportions are desired.}
}
\value{
A nested \code{tibble} with special class \code{category}.
}
\description{
This function creates an individual weighting category with known marginal 
probabilities (E.g., age group, eye color.), but unlike \code{category()} it 
creates them from existing data instead of being directly assigned. One or more 
of these are built and fed into \code{universe()}.
}
\examples{
data(weight_me)

inherit_category(
  name = "costume",
  from = dplyr::filter(weight_me, satisfied == "No")
)

inherit_category(
  name = "transport",
  from = dplyr::filter(weight_me, satisfied == "No"),
  weight = prev_weight
)

}
