% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_category.R
\name{new_category}
\alias{new_category}
\title{Append universe with new category}
\usage{
new_category(universe, name, groups, targets)
}
\arguments{
\item{universe}{Output object from \code{universe()}.}

\item{name}{The name of the weighting category matching the name of the
corresponding variable in the data to be weighted.}

\item{groups}{A vector of the groups in the weighting category.}

\item{targets}{A vector of the desired proportions of the \code{groups}.
These targets will be automatically re-proportioned to sum to 1 if
necessary.}
}
\value{
A \code{list} with special class \code{universe}.
}
\description{
This appends an existing \code{universe} object with a new weighting category.
}
\examples{
uni <- universe(
    data = mtcars,

    category(
        name = "vs",
        groups = c(0, 1),
        targets = c(1/2, 1/2)
    )
)

new_category(
    universe = uni, 
    name = "am",
    groups = c(0, 1),
    targets = c(1/3, 2/3)
)

}
