% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_model.R
\name{pop_model}
\alias{pop_model}
\title{Create full weighting design}
\usage{
pop_model(df, ...)
}
\arguments{
\item{df}{Data frame containing data you intend to weight.}

\item{...}{Object or objects of special class \code{wgt_cat}.}
}
\value{
A nested \code{tibble} with special class \code{pop_model}.
}
\description{
This combines any number of objects with special class \code{wgt_cat}. It also checks and adjusts
for \code{NA} data in the data frame to be weighted.
}
\examples{
data("weight_me")

pop_model(
    df = weight_me,

    wgt_cat(
        name = "gender",
        buckets = c(1, 2),
        targets = c(0.5, 0.5)),

    wgt_cat(
        name = "vehicle",
        buckets = c(1, 2, 3),
        targets = c(0.333, 0.333, 0.333))
)

}
