% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_rake.R
\name{pre_rake}
\alias{pre_rake}
\title{Compare unweighted data to targets}
\usage{
pre_rake(df, design, plot = FALSE)
}
\arguments{
\item{df}{Data frame of the data you intend on weighting.}

\item{plot}{Display plot, default = FALSE.}

\item{pop.model}{Full weighting design created by \code{wgt_design()}.}
}
\value{
A tibble of unweighted counts and proportions, difference between 
unweighted and target proportions. Optionally, a plot of this information.
}
\description{
Inspect by what degree the unweighted sample proportions deviate from the
target proportions in the population model.
}
\examples{
data("weight_me")

mod <- wgt_design(

    df = weight_me,

    wgt_cat(
        name = "gender",
        buckets = c(1, 2),
        targets = c(0.5, 0.5)),

    wgt_cat(
        name = "vehicle",
        buckets = c(1, 2, 3),
        targets = c(0.333, 0.333, 0.333))
)

pre_rake(
    df = weight_me,
    pop.model = mod,
    plot = TRUE
)

}
