% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{print.category}
\alias{print.category}
\title{Print method for category objects.}
\usage{
\method{print}{category}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{category} object.}

\item{digits}{Number of digits for printing the target proportions, default is 3.}

\item{...}{Not currently used.}
}
\description{
Print method for category objects.
}
