% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{print.universe}
\alias{print.universe}
\title{Print method for universe objects.}
\usage{
\method{print}{universe}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{universe} object.}

\item{digits}{Number of digits for printing the target proportions, default is 3.}

\item{...}{Not currently used.}
}
\description{
Print method for universe objects.
}
