% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_size}
\alias{sample_size}
\title{Sample size calculations.}
\usage{
sample_size(x, w, type = "u")
}
\arguments{
\item{x}{A vector.}

\item{w}{A vector of weights.}

\item{type}{One of \code{"u"} for unweighted, \code{"w"} for weighted, \code{"e"} effective,
or \code{"all"} for all three.}
}
\value{
Either a single numeric value, or a named vector if \code{type = "all"}.
}
\description{
The unweighed sample size is the number of non-missing elements in \code{x}. The
weighted sample size is the sum of the weights for the complete cases in \code{x}
and \code{w}. The effective sample size is \eqn{(sum(w)^2) / sum(w^2)}.
}
\examples{
x <- 1:3
w <- c(0.75, 1.00, 1.25)

sample_size(x, w, type = "u")
sample_size(x, w, type = "w")
sample_size(x, w, type = "e")
sample_size(x, w, type = "all")

}
