% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe}
\alias{universe}
\title{Create universe}
\usage{
universe(data, ..., N)
}
\arguments{
\item{data}{Data frame containing data where weights are desired.}

\item{...}{One or more output objects from \code{category()}.}

\item{N}{Size of universe. If supplied, expansion factor will be applied
to weights after convergence.}
}
\value{
A \code{list} with special class \code{universe}.
}
\description{
Along with the data to be weighted, this combines one or more objects of 
class \code{category} to build the universe with known marginal proportions. 
It also checks and adjusts the proportions to account for missing values in the data.
}
\examples{
data(demo_data)

universe(
    data = demo_data,
    
    category(
        name = "Sex",
        buckets = factor(
            x = levels(demo_data[["Sex"]]),
            levels = levels(demo_data[["Sex"]])
        ),
        targets = c(0.4, 0.5),
        sum.1 = TRUE
    ),

    category(
        name = "BirthYear",
        buckets = c(1986:1990),
        targets = rep(0.2, times = 5)
    ),
    
    category(
        name = "EyeColor",
        buckets = c("brown", "green", "blue"),
        targets = c(0.8, 0.1, 0.1)
    ),
    
    category(
        name = "HomeOwner",
        buckets = c(TRUE, FALSE),
        targets = c(3/4, 1/4)
    )
)

}
