% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unweighted_ss}
\alias{unweighted_ss}
\title{Unweighted sample size.}
\usage{
unweighted_ss(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Unweighted sample size.
}
\examples{
unweighted_ss(
    x = runif(3)
)

}
