% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_factors.R
\name{weight_factors}
\alias{weight_factors}
\title{Get weighting factors for categories used in weighting scheme}
\usage{
weight_factors(x)
}
\arguments{
\item{x}{Output of \code{iterake()}.}
}
\value{
A \code{tibble}.
}
\description{
This creates a \code{tibble} of every combination of the groups of the weighting
categories passed to \code{universe()}, their weights, and the number of occurrences.
}
