% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats.iterake}
\alias{weight_stats.iterake}
\title{Weight summary statistics for an \code{iterake} object}
\usage{
\method{weight_stats}{iterake}(x)
}
\arguments{
\item{x}{Output of \code{iterake()}.}
}
\description{
Weight summary statistics for an \code{iterake} object
}
