% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats.numeric}
\alias{weight_stats.numeric}
\title{Weight summary statistics for a numeric vector}
\usage{
\method{weight_stats}{numeric}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Weight summary statistics for a numeric vector
}
