% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats}
\alias{weight_stats}
\title{Generate summary statistics for a set of weights}
\usage{
weight_stats(weight)
}
\arguments{
\item{weight}{Single column of a data frame or vector of weights.}
}
\value{
A tibble containing unweighted and weighted sample size, effective sample size, and
efficiency and loss values for the supplied weight data.
}
\description{
This function creates a tibble containing information pertaining to a set of weights. It includes
sample size, weighted sample size, effective sample size, as well as loss and efficiency values.
}
\examples{
data(demo_data)

wgts <- 
  iterake(
    universe = universe(
        data = demo_data,
        
        category(
            name = "Sex",
            buckets = factor(
                x = levels(demo_data[["Sex"]]),
                levels = levels(demo_data[["Sex"]])
            ),
            targets = c(0.4, 0.5),
            sum.1 = TRUE
        ),

        category(
            name = "BirthYear",
            buckets = c(1986:1990),
            targets = rep(0.2, times = 5)
        ),
    
        category(
            name = "EyeColor",
            buckets = c("brown", "green", "blue"),
            targets = c(0.8, 0.1, 0.1)
        ),
    
        category(
            name = "HomeOwner",
            buckets = c(TRUE, FALSE),
            targets = c(3/4, 1/4)
        )
    )
  )

weight_stats(wgts$weight)

}
