% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_ss}
\alias{weighted_ss}
\title{Weighted sample size.}
\usage{
weighted_ss(x, w)
}
\arguments{
\item{x}{A vector.}

\item{w}{A vector of weights.}
}
\description{
Weighted sample size.
}
\examples{
weighted_ss(
    x = runif(3),
    w = c(0.75, 1.00, 1.25)
)

}
