% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighting_efficiency}
\alias{weighting_efficiency}
\title{Weighting efficiency.}
\usage{
weighting_efficiency(w)
}
\arguments{
\item{w}{A vector of weights.}
}
\description{
Weighting efficiency.
}
\examples{
weighting_efficiency(c(0.75, 1.00, 1.25))

}
