% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_cat_inherit.R
\name{wgt_cat_inherit}
\alias{wgt_cat_inherit}
\title{Create weighting category from data}
\usage{
wgt_cat_inherit(name, df, prev.wgt)
}
\arguments{
\item{name}{Name given to weighting category, character. Must have exact match in the column names of data you intend to weight.}

\item{df}{Data frame containing data you intend to weight.}

\item{prev.wgt}{Variable name of weight used to calculate bucket proportions in \code{`df`}.}
}
\value{
A nested \code{tibble} with special class \code{wgt_cat}.
}
\description{
To be fed into \code{pop.model()}, this function creates an individual weighting category. (E.g., age group.)
Unlike \code{wgt_cat()}, this function creates bucket targets from a dataframe instead of being directly assigned.
}
\examples{
data("weight_me")

wgt_cat_inherit(
  name = "gender",
  df = weight_me \%>\% filter(group == 2)
)

wgt_cat_inherit(
  name = "vehicle",
  df = weight_me \%>\% filter(group == 2),
  prev.wgt = origWeight
)

}
