% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_design.R
\name{wgt_design}
\alias{wgt_design}
\title{Create full weighting design}
\usage{
wgt_design(df, ...)
}
\arguments{
\item{df}{Data frame containing data you intend to weight.}

\item{...}{Object or objects of special class \code{wgt_cat}.}
}
\value{
A nested \code{tibble} with special class \code{wgt_design}.
}
\description{
This combines any number of objects with special class \code{wgt_cat}. It also checks and adjusts
for \code{NA} data in the data frame to be weighted.
}
\examples{
data(weight_me)

wgt_design(
    df = weight_me,

    wgt_cat(
        name = "costume",
        buckets = c("Bat Man", "Cactus"),
        targets = c(0.5, 0.5)),

    wgt_cat(
        name = "seeds",
        buckets = c("Tornado", "Bird", "Earthquake"),
        targets = c(0.4, 0.3, 0.3))
)

}
