% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_review.R
\name{wgt_review}
\alias{wgt_review}
\title{Compare unweighted or weighted data to targets}
\usage{
wgt_review(df, design, weight, plot = FALSE)
}
\arguments{
\item{df}{Data frame of the data you intend on weighting.}

\item{design}{Full weighting design created by \code{wgt_design()}.}

\item{weight}{Name of weight variable, optional.}

\item{plot}{Display plot, default = FALSE.}
}
\value{
A tibble of unweighted counts and proportions, difference between 
unweighted and target proportions. If \code{`weight`} is given, it also includes
weighted counts and proportions, difference between weighted and target proportions.
Optionally, a plot of this information.
}
\description{
Inspect by what degree the unweighted or weighted sample proportions deviate from the
target proportions in the weight design.
}
\examples{
data(weight_me)

mod <- wgt_design(

    df = weight_me,

    wgt_cat(
        name = "costume",
        buckets = c("Bat Man", "Cactus"),
        targets = c(0.5, 0.5)),

    wgt_cat(
        name = "seeds",
        buckets = c("Tornado", "Bird", "Earthquake"),
        targets = c(0.3, 0.3, 0.4))
)

wgt_review(
    df = weight_me,
    design = mod,
    plot = TRUE
)

wgts <- iterake(
    df = weight_me,
    id = order, 
    design = mod
)

wgt_review(
    df = wgts,
    design = mod,
    weight = weight,
    plot = FALSE)

}
