% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{category}
\alias{category}
\title{Create a weighting category.}
\usage{
category(name, groups, targets)
}
\arguments{
\item{name}{The name of the weighting category matching the name of the
corresponding variable in the data to be weighted.}

\item{groups}{A vector of the groups in the weighting category.}

\item{targets}{A vector of the desired proportions of the \code{groups}.
These target will be automatically re-proportioned to sum to 1 if
necessary.}
}
\value{
A list with the class of \code{"category"}.
}
\description{
This function creates a new weighting category.
}
\examples{
category(
    name = "vs",
    groups = c(0, 1),
    targets = c(1/3, 2/3)
)

}
