% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_margins.R
\name{compare_margins}
\alias{compare_margins}
\title{Compare unweighted or weighted data to targets}
\usage{
compare_margins(df, weight, universe, plot = FALSE)
}
\arguments{
\item{df}{Data frame of the data you intend on weighting.}

\item{weight}{Name of weight variable, optional.}

\item{universe}{Full weighting design created by \code{universe()}.}

\item{plot}{Display plot, default = FALSE.}
}
\value{
A tibble of unweighted counts and proportions, difference between 
unweighted and target proportions. If \code{`weight`} is given, it also includes
weighted counts and proportions, difference between weighted and target proportions.
Optionally, a plot of this information.
}
\description{
Inspect by what degree the unweighted or weighted sample proportions deviate from the
target proportions.
}
\examples{
data(weight_me)

mod <- universe(

    df = weight_me,

    category(
        name = "costume",
        buckets = c("Bat Man", "Cactus"),
        targets = c(0.5, 0.5)),

    category(
        name = "seeds",
        buckets = c("Tornado", "Bird", "Earthquake"),
        targets = c(0.3, 0.3, 0.4))
)

compare_margins(
    df = weight_me,
    universe = mod,
    plot = TRUE
)

wgts <- iterake(
    df = weight_me,
    universe = mod
)

compare_margins(
    df = wgts,
    universe = mod,
    weight = weight,
    plot = FALSE)

}
