% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterake.R
\name{iterake}
\alias{iterake}
\title{Iterative raking procedure}
\usage{
iterake(universe, permute = FALSE, control = control_iterake())
}
\arguments{
\item{universe}{Output object created with \code{universe()} function.}

\item{permute}{Boolean indicating whether to test all possible orders of categories in \code{universe}
and keep the most efficient (\code{TRUE}) or to test categories in the order listed in \code{universe}
only (default, \code{FALSE}), optional. Note that when \code{TRUE} this will increase runtime by a
factor of \code{(number of categories)!}.}

\item{control}{Output object created with \code{control_iterake()} function.}
}
\value{
A \code{list} that includes details of the run as well as the generated weights.
}
\description{
This function creates case weights using an iterative raking algorithm based
on targets from a known population (established with \code{universe()}).
The weights are appended as a new column in the data. If \code{iterake()}
converges, the weighted marginal proportions of the sample will match those
set in \code{universe()}. Summary statistics of the weighting procedure are
presented by default.
}
