% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterake.R
\name{iterake}
\alias{iterake}
\title{Iterative raking procedure}
\usage{
iterake(df, universe, wgt.name = "weight", max.wgt = 3,
  threshold = 1e-20, max.iter = 50, stuck.limit = 5,
  summary = TRUE)
}
\arguments{
\item{df}{Data frame containing data where weights are desired.}

\item{universe}{Output object created with \code{universe()} function.}

\item{wgt.name}{Name given to column of weights to be added to data, optional.}

\item{max.wgt}{Maximum value weights can take on, optional.}

\item{threshold}{Value specifying minimum summed difference between weighted 
marginal proportions of sample and universe before algorithm quits, optional.}

\item{max.iter}{Value capping number of iterations for the procedure, optional.}

\item{stuck.limit}{Value capping the number of times summed differences between 
sample and universe can oscillate between increasing and decreasing, optional.}

\item{summary}{Whether or not to display summary output of the procedure, optional.}
}
\value{
Data frame with the resulting weight variable appended to it.
}
\description{
This function creates row-level weights using an iterative raking algorithm based
on targets from a known population (established with \code{universe()}).
The weights are appended as a new column in the data. If \code{iterake()} 
converges, the weighted marginal proportions of the sample will match those
set in \code{universe()}. Summary statistics of the weighting procedure are 
presented by default.
}
\examples{
library(iterake)
data(weight_me)

iterake(
    df = weight_me,
    universe = universe(
        df = weight_me,

        build_margin(
            name = "seeds",
            buckets = c("Tornado", "Bird", "Earthquake"),
            targets = c(0.300, 0.360, 0.340)),

        build_margin(
            name = "costume",
            buckets = c("Bat Man", "Cactus"),
            targets = c(0.500, 0.500)),

        build_margin(
            name = "transport",
            buckets = c("Rocket Cart", "Jet Propelled Skis", "Jet Propelled Unicycle"),
            targets = c(0.400, 0.450, 0.150))

    )
)

}
