% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterake.R
\name{iterake}
\alias{iterake}
\title{Iterative rim weighting procedure}
\usage{
iterake(df, id, design, wgt.name = "weight", wgt.lim = 3,
  threshold = 1e-20, max.iter = 50, stuck.limit = 5, N, summary = TRUE)
}
\arguments{
\item{df}{Data frame containing data you intend to weight.}

\item{id}{Variable name for unique identifier in \code{`df`}.}

\item{design}{Data frame containing data you intend to weight.}

\item{wgt.name}{Character name for created weight variable, optional.}

\item{wgt.lim}{Numeric value created weights cannot exceed, optional.}

\item{threshold}{Numeric value specifying minimum summed difference between target and weighted proportions, optional.}

\item{max.iter}{Numeric value capping number of iterations for the procedure, optional.}

\item{stuck.limit}{Numeric value capping the number of times summed differences between target and weighted
proportions can oscillate between increasing and decreasing, optional.}

\item{N}{Numeric value representing expansion factor to be applied to generated weights, optional.}

\item{summary}{Boolean value for whether to show summary output of the procedure, optional.}
}
\value{
Data frame with the resulting weight variable appended to it.
}
\description{
This function creates a weight variable that satisfies the targets described from the output of \code{wgt_design()}.
It then appends the weight variable to the data frame used to create the weights.
}
\examples{
data(weight_me)

iterake(
    df = weight_me,
    id = order, 
    design = wgt_design(
        df = weight_me,

        wgt_cat(
            name = "seeds",
            buckets = c("Tornado", "Bird", "Earthquake"),
            targets = c(0.300, 0.360, 0.340)),

        wgt_cat(
            name = "costume",
            buckets = c("Bat Man", "Cactus"),
            targets = c(0.500, 0.500)),

        wgt_cat(
            name = "transport",
            buckets = c("Rocket Cart", "Jet Propelled Skis", "Jet Propelled Unicycle"),
            targets = c(0.400, 0.450, 0.150))

    )
)

}
