% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_table.R
\name{label_table}
\alias{label_table}
\title{Create Data Frame of Variable or Value Labels.}
\usage{
label_table(data, var, freq = TRUE, weight)
}
\arguments{
\item{data}{A data frame where variable or value labels are requested.}

\item{var}{If value labels are requested, provide corresponding column (unquoted) in the data frame.}

\item{freq}{Logical, whether or not to tabulate frequencies for the value labels.}

\item{weight}{An unquoted column of weights.}
}
\value{
A data frame of variable or value labels.
}
\description{
This enables you to inspect labels at the variable or value label.
If requesting value labels for a \code{var}, the default is to also
tabulate the frequencies and percents. Can provide an optional column
of weights to calculated weighted frequencies and percentages.
}
\examples{
library(dplyr)
library(labelled)
library(iterake)

df <-
  data.frame(
    x = sample(c(1, 2), 100, replace = TRUE),
    w = rnorm(100, 1, 0.1)
  ) \%>\%
  set_variable_labels(
    x = "Gender",
    w = "Weight"
  ) \%>\%
  set_value_labels(
    x = c("Male" = 1, "Female" = 2)
  )

label_table(df)
label_table(df, x, freq = FALSE)
label_table(df, x)
label_table(df, x, weight = w)

}
