% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_table.R
\name{label_table}
\alias{label_table}
\title{Create table of variable or value labels}
\usage{
label_table(data, var, freq = FALSE, weight)
}
\arguments{
\item{data}{Data frame.}

\item{var}{A variable in the dataset where labels are desired.}

\item{freq}{Boolean, indicates whether frequencies should be included, optional.}

\item{weight}{A variable in the dataset representing weights to be used 
for optional frequency calculation, optional.}
}
\value{
By default, will return a dataset of variables and labels of every
variable in the dataset. If specified, returns a table of values in a
variable and its labels, as well as frequencies.
}
\description{
It is often useful to inspect variable or value labels attached to datasets
(like with SAV files from SPSS). This function will return a tibble of variables and
their associated labels, or, if requested, a variable and its value labels.
}
\examples{
data(weight_me)

label_table(data = weight_me, var = seeds)

}
