% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterake.R
\name{print.iterake}
\alias{print.iterake}
\title{Print method for iterake objects.}
\usage{
\method{print}{iterake}(x, digits = 3, ...)
}
\arguments{
\item{x}{An \code{iterake} object.}

\item{digits}{Number of digits for printing proportions, default is 3.}

\item{...}{Not currently used.}
}
\description{
Print method for iterake objects.
}
