% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe}
\alias{universe}
\title{Create universe}
\usage{
universe(df, ...)
}
\arguments{
\item{df}{Data frame containing data where weights are desired.}

\item{...}{One or more output objects from \code{category()}.}
}
\value{
A nested \code{tibble} with special class \code{universe}.
}
\description{
Along with the data to be weighted, this combines one or more objects of 
class \code{category} to build the universe with known marginal proportions. 
It also checks and adjusts the proportions to account for missing values in the data.
}
\examples{
library(iterake)

data(weight_me)

universe(
    df = weight_me,

    category(
        name = "costume",
        buckets = c("Bat Man", "Cactus"),
        targets = c(0.5, 0.5)),

    category(
        name = "seeds",
        buckets = c("Tornado", "Bird", "Earthquake"),
        targets = c(0.4, 0.3, 0.3))
)

}
