% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe}
\alias{universe}
\title{Create universe}
\usage{
universe(data, ...)
}
\arguments{
\item{data}{Data frame containing data where weights are desired.}

\item{...}{One or more output objects from \code{category()}.}
}
\value{
A \code{list} with special class \code{universe}.
}
\description{
Along with the data to be weighted, this combines one or more objects of
class \code{category} to build the universe with known marginal proportions.
It also checks and adjusts the proportions to account for missing values in the data.
}
\examples{
universe(
    data = mtcars,
    category(
        name = "vs",
        groups = c(0, 1),
        targets = c(1/2, 1/2)
    )
)

}
