% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe}
\alias{universe}
\title{Create full weighting design}
\usage{
universe(df, ...)
}
\arguments{
\item{df}{Data frame containing data you intend to weight.}

\item{...}{Object or objects of special class \code{build_margin}.}
}
\value{
A nested \code{tibble} with special class \code{universe}.
}
\description{
This combines any number of objects with special class \code{build_margin}. It also checks and adjusts
for \code{NA} data in the data frame to be weighted.
}
\examples{
data(weight_me)

universe(
    df = weight_me,

    build_margin(
        name = "costume",
        buckets = c("Bat Man", "Cactus"),
        targets = c(0.5, 0.5)),

    build_margin(
        name = "seeds",
        buckets = c("Tornado", "Bird", "Earthquake"),
        targets = c(0.4, 0.3, 0.3))
)

}
