% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_factors.R
\name{weight_factors}
\alias{weight_factors}
\title{Generate information on weight factors}
\usage{
weight_factors(x)
}
\arguments{
\item{x}{Output of \code{iterake()}.}
}
\value{
A \code{tibble} containing every weighting category combination used by \code{iterake()} along with
the weight assigned to it and the number in each combination.
}
\description{
This function creates a \code{tibble} containing every combination of weighting categories
used by \code{iterake()}, along with the size of each combination and its weight.
}
