% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_factors.R
\name{weight_factors}
\alias{weight_factors}
\title{Generate information on the weight factors used by iterake.}
\usage{
weight_factors(iterake)
}
\arguments{
\item{iterake}{Output object created with \code{iterake()} function.}
}
\value{
A tibble containing every weighting category combination used by iterake along with
the weight assigned to it and the number in each combination.
}
\description{
This function creates a tibble containing every combination of weighting categories
used by iterake, along with the size of each combination and its weight.
}
