% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats}
\alias{weight_stats}
\title{Generate summary statistics for generated weights}
\usage{
weight_stats(iterake)
}
\arguments{
\item{iterake}{Output object created with \code{iterake()} function.}
}
\value{
A tibble containing unweighted and weighted sample size, effective sample size, and
efficiency and loss values for the weights generated by iterake.
}
\description{
This function creates a tibble containing information pertaining to a set of weights. It includes
sample size, weighted sample size, effective sample size, as well as loss and efficiency values.
}
