% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats}
\alias{weight_stats}
\title{Weight summary statistics}
\usage{
weight_stats(x)
}
\arguments{
\item{x}{An \code{iterake} object or a numeric vector.}
}
\value{
A \code{tibble}.
}
\description{
Returns a \code{tibble} of the unweighted, weighted, and effective sample sizes,
the loss and weighting efficiency, mean and median of the weight,
and the minimum and maximum weights.
}
