% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats}
\alias{weight_stats}
\title{Summary weight statistics}
\usage{
weight_stats(x)
}
\arguments{
\item{x}{Output of \code{iterake()}.}
}
\value{
A \code{tibble} containing unweighted and weighted sample size, effective sample size, and
efficiency and loss values for the weights generated by \code{iterake()}.
}
\description{
This function creates a \code{tibble} containing information pertaining to a set of weights. It includes
sample size, weighted sample size, effective sample size, as well as loss and efficiency values.
}
