% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_stats.R
\name{weight_stats}
\alias{weight_stats}
\title{Generate summary statistics for a set of weights}
\usage{
weight_stats(weight)
}
\arguments{
\item{weight}{Single column of a data frame or vector of weights.}
}
\value{
A tibble containing unweighted and weighted sample size, effective sample size, and
efficiency and loss values for the supplied weight data.
}
\description{
This function creates a tibble containing information pertaining to a set of weights. It includes
sample size, weighted sample size, effective sample size, as well as loss and efficiency values.
}
\examples{
data(weight_me)

wgts <- iterake(
    df = weight_me,
    universe = universe(
        df = weight_me,

        category(
            name = "seeds",
            buckets = c("Tornado", "Bird", "Earthquake"),
            targets = c(0.300, 0.360, 0.340)),

        category(
            name = "costume",
            buckets = c("Bat Man", "Cactus"),
            targets = c(0.500, 0.500)),

        category(
            name = "transport",
            buckets = c("Rocket Cart", "Jet Propelled Skis", "Jet Propelled Unicycle"),
            targets = c(0.400, 0.450, 0.150))

    )
)

weight_stats(wgts$weight)

}
