% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_review.R
\name{wgt_review}
\alias{wgt_review}
\title{Compare unweighted or weighted data to targets}
\usage{
wgt_review(df, design, weight, plot = FALSE)
}
\arguments{
\item{df}{Data frame of the data you intend on weighting.}

\item{design}{Full weighting design created by \code{wgt_design()}.}

\item{weight}{Name of weight variable, optional.}

\item{plot}{Display plot, default = FALSE.}
}
\value{
A tibble of unweighted counts and proportions, difference between 
unweighted and target proportions. If \code{`weight`} is given, it also includes
weighted counts and proportions, difference between weighted and target proportions.
Optionally, a plot of this information.
}
\description{
Inspect by what degree the unweighted or weighted sample proportions deviate from the
target proportions in the weight design.
}
\examples{
data("weight_me")

mod <- wgt_design(

    df = weight_me,

    wgt_cat(
        name = "gender",
        buckets = c(1, 2),
        targets = c(0.5, 0.5)),

    wgt_cat(
        name = "vehicle",
        buckets = c(1, 2, 3),
        targets = c(0.333, 0.333, 0.333))
)

wgt_review(
    df = weight_me,
    design = mod,
    plot = TRUE
)

}
