% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_stats.R
\name{wgt_stats}
\alias{wgt_stats}
\title{Generate summary statistics for a set of weights}
\usage{
wgt_stats(weight)
}
\arguments{
\item{weight}{Single column of a data frame or vector of weights.}
}
\value{
A tibble containing unweighted and weighted sample size, effective sample size, and
efficiency and loss values for the supplied weight data.
}
\description{
This function creates a tibble containing information pertaining to a set of weights. It includes
sample size, weighted sample size, effective sample size, as well as loss and efficiency values.
}
\examples{
data("weight_me")

wgts <- iterake(
    df = weight_me,
    id = id, 
    design = wgt_design(df = weight_me,

                        wgt_cat(name = "age",
                                buckets = c("18-34", "35-54", "55+"),
                                targets = c(0.300, 0.360, 0.340)),

                        wgt_cat(name = "gender",
                                buckets = c("Female", "Male"),
                                targets = c(0.500, 0.500)),

                        wgt_cat(name = "vehicle",
                                buckets = c("Car", "SUV", "Truck"),
                                targets = c(0.400, 0.450, 0.150))

    )
)

wgt_stats(wgts$weight)

}
